myblog/requirements.txt
4==8.8.8.8
asgiref==3.6.0
attrs==23.1.0
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
drf-spectacular==0.26.2
exceptiongroup==1.1.1
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
Markdown==3.4.3
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
tomli==2.0.1
uritemplate==4.1.1
